+++ apps/docs/content/landing/index.tsx
@@ -304,7 +304,7 @@ export const MusicPlayer: FC<MusicPlayerProps> = ({className, ...otherProps}) =>
   return (
     <Card
       isBlurred
-      className={clsx("border-none dark:bg-background/40 bg-background/40", className)}
+      className={clsx("border-none bg-background/60 dark:bg-default-100/50", className)}
       radius="2xl"
       shadow="2xl"
       {...otherProps}
@@ -350,7 +350,7 @@ export const MusicPlayer: FC<MusicPlayerProps> = ({className, ...otherProps}) =>
               <Progress
                 aria-label="Music progress"
                 classNames={{
-                  filler: "bg-white",
+                  indicator: "bg-default-800 dark:bg-white",
                   track: "bg-default-500/30",
                 }}
                 color="default"