+++ packages/components/chip/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nextui-org/chip
 
+## 0.0.0-dev-v2-20230625145649
+
+### Patch Changes
+
+- New color added to manage overlays, popover and dropdown styles small changes
+- Updated dependencies
+  - @nextui-org/shared-icons@0.0.0-dev-v2-20230625145649
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230625145649
+  - @nextui-org/react-utils@0.0.0-dev-v2-20230625145649
+  - @nextui-org/system@0.0.0-dev-v2-20230625145649
+  - @nextui-org/theme@0.0.0-dev-v2-20230625145649
+
 ## 0.0.0-dev-v2-20230625035223
 
 ### Patch Changes